more 64 bit fixes
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19133 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/hashhandlers.h
@@ -86,7 +86,7 @@ class csRefHashKeyHandler
 public:
   static uint32 ComputeHash (const csRef<T>& key)
   {
-    return (uint32)((T*)key);
+    return (intptr_t) ((T*)key);
   }
 
   static bool CompareKeys (const csRef<T>& key1, const csRef<T>& key2)
